Revert "删除了DTO，全部使用POJO来接收并校验前端传递的数据，返回的数据统一使用VO包装。 引入Kotlin，往下打算采用Kotlin全部重写整个项目"
This reverts commit e86479117b8b0e726e698f18fade38fce4b74603.

+++ src/main/java/com/fish/service/impl/LabelServiceImpl.java
@@ -9,22 +9,22 @@ import com.fish.service.LabelService;
 import com.fish.utils.ResultUtil;
 import com.mybatisflex.spring.service.impl.ServiceImpl;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.StringUtils;
 
 import java.util.ArrayList;
 import java.util.Objects;
 
 @Service
 public class LabelServiceImpl extends ServiceImpl<LabelMapper, Label> implements LabelService {
-    @Transactional
     @Override
     public Result<?> addLabel(Label label) {
+        if (!StringUtils.hasLength(label.getLabelName()))
+            throw new ServiceException(ServiceExceptionEnum.KEY_ARGUMENT_NOT_INPUT);
         int i = mapper.insert(label);
         if (i > 0)
             return ResultUtil.success();
         throw new ServiceException(ServiceExceptionEnum.OPERATE_ERROR);
     }
-    @Transactional
     @Override
     public Result<?> addLabelBatch(ArrayList<Label> labels) {
         if (labels.isEmpty())
@@ -41,22 +41,26 @@ public class LabelServiceImpl extends ServiceImpl<LabelMapper, Label> implements
     }
     @Override
     public Result<Label> getLabel(Long labelId) {
+        if (Objects.isNull(labelId))
+            throw new ServiceException(ServiceExceptionEnum.KEY_ARGUMENT_NOT_INPUT);
         Label label = mapper.selectOneById(labelId);
         if (!Objects.isNull(label))
             return ResultUtil.success(label);
         throw new ServiceException(ServiceExceptionEnum.NOT_FOUND);
     }
-    @Transactional
     @Override
     public Result<?> updateLabel(Label label) {
+        if (!StringUtils.hasLength(label.getLabelName()) || Objects.isNull(label.getLabelId()))
+            throw new ServiceException(ServiceExceptionEnum.KEY_ARGUMENT_NOT_INPUT);
         int i = mapper.update(label);
         if (i > 0)
             return ResultUtil.success();
         throw new ServiceException(ServiceExceptionEnum.OPERATE_ERROR);
     }
-    @Transactional
     @Override
     public Result<?> deleteLabel(Long labelId) {
+        if (Objects.isNull(labelId))
+            throw new ServiceException(ServiceExceptionEnum.KEY_ARGUMENT_NOT_INPUT);
         int i = mapper.deleteById(labelId);
         if (i > 0)
             return ResultUtil.success();
